pyproject.toml
[project]
name = "phrase-drawer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "rembg>=2.0.67",
    "scipy>=1.16.3",
]
